Finalize workspace modularization and ungate bulk edit
- Split the workspace template into focused Thymeleaf fragments
and keep workspace.html as the thin composition shell.

- Remove the remaining bulk edit feature gating so the button
and modal are always rendered and the frontend always binds.

- Refresh workspace render coverage and update .context docs
to reflect the delivered modular frontend structure.
