+++ src/main/resources/application.properties
@@ -12,9 +12,5 @@ spring.datasource.username=${DB_USERNAME}
 spring.datasource.password=${DB_PASSWORD}
 spring.datasource.driver-class-name=org.postgresql.Driver
 
-# Ignore editor warnings, these flags are implemented :) 
-# Bulk edit rollout controls.
-# Keep disabled until search scalability work is deployed.
-teststream.bulk-edit.enabled=true
 # Safety cap for number of requested work keys in one bulk edit call.
 teststream.bulk-edit.max-work-keys=5000
