+++ src/main/resources/templates/workspace/_header.html
@@ -1,0 +1,143 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<body>
+	<th:block th:fragment="header">
+		<header class="border-b border-white/10">
+			<div class="w-full px-6 sm:px-8 py-5 flex flex-col lg:flex-row lg:items-center lg:justify-between gap-4">
+				<div class="flex-1 min-w-0">
+					<div class="flex flex-col sm:flex-row sm:items-center gap-3">
+						<div class="flex-1 min-w-0">
+							<label class="block text-xs uppercase tracking-wider text-white/45" style="font-family: 'Work Sans', sans-serif; font-weight: 700;">Search</label>
+							<div class="mt-2">
+								<input id="wsSearch" type="text"
+									placeholder="Search ID, title, tags..."
+									class="w-full bg-black border border-white/15 px-4 py-3 text-sm text-white/85 focus:outline-none focus:border-[#E7FF02]" />
+							</div>
+						</div>
+
+						<div class="grid grid-cols-1 sm:grid-cols-3 gap-3 sm:w-[34rem]">
+							<div>
+								<label class="block text-xs uppercase tracking-wider text-white/45" style="font-family: 'Work Sans', sans-serif; font-weight: 700;">Component</label>
+								<select id="wsFilterComponent" class="mt-2 w-full bg-black border border-white/15 px-3 py-3 text-sm text-white/85 focus:outline-none focus:border-[#E7FF02]">
+									<option value="">All</option>
+								</select>
+							</div>
+							<div>
+								<label class="block text-xs uppercase tracking-wider text-white/45" style="font-family: 'Work Sans', sans-serif; font-weight: 700;">Status</label>
+								<select id="wsFilterStatus" class="mt-2 w-full bg-black border border-white/15 px-3 py-3 text-sm text-white/85 focus:outline-none focus:border-[#E7FF02]">
+									<option value="">All</option>
+								</select>
+							</div>
+							<div>
+								<label class="block text-xs uppercase tracking-wider text-white/45" style="font-family: 'Work Sans', sans-serif; font-weight: 700;">Tags</label>
+								<select id="wsFilterTag" class="mt-2 w-full bg-black border border-white/15 px-3 py-3 text-sm text-white/85 focus:outline-none focus:border-[#E7FF02]">
+									<option value="">All</option>
+								</select>
+							</div>
+						</div>
+					</div>
+				</div>
+
+				<div class="shrink-0 flex flex-wrap items-start justify-between lg:justify-end gap-3 lg:pt-6">
+					<div class="flex items-center gap-2">
+						<form id="importForm" th:action="@{/workspace/import}" method="post" enctype="multipart/form-data" class="flex items-center">
+							<input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}" />
+							<input id="importFile" type="file" name="file" accept=".xlsx,.csv" class="hidden" />
+							<button type="button" id="importBtn" class="px-4 py-2 border border-white/20 hover:border-[#E7FF02] hover:text-[#E7FF02] transition-colors text-sm">
+								Import
+							</button>
+						</form>
+						<a th:href="@{/workspace/export}" class="px-4 py-2 border border-white/20 hover:border-[#E7FF02] hover:text-[#E7FF02] transition-colors text-sm">
+							Export
+						</a>
+					</div>
+
+					<div class="flex items-center gap-3">
+						<div class="hidden sm:flex items-start gap-4">
+							<div class="relative text-right min-w-0">
+								<button
+									type="button"
+									class="text-xs uppercase tracking-wider text-white/45 hover:text-[#E7FF02] transition-colors select-none"
+									style="font-family: 'Work Sans', sans-serif; font-weight: 700;"
+									data-workspace-dropdown-toggle="account"
+									aria-haspopup="true"
+									aria-controls="workspaceAccountDropdown"
+									aria-expanded="false"
+									id="workspaceAccountToggle"
+								>
+									Account <span aria-hidden="true" class="text-white/45"></span>
+								</button>
+								<div
+									id="workspaceAccountDropdown"
+									class="hidden absolute right-0 mt-2 z-50 w-64 border border-white/20 bg-black px-4 py-3 text-left"
+									role="dialog"
+									aria-label="Account details"
+									aria-hidden="true"
+								>
+									<p class="text-xs uppercase tracking-wider text-white/45" style="font-family: 'Work Sans', sans-serif; font-weight: 700;">Signed in as</p>
+									<p class="mt-1 text-sm text-white/80 break-words" th:text="${userEmail != null ? userEmail : 'Signed in'}">Signed in</p>
+								</div>
+							</div>
+
+							<div class="relative text-right min-w-0" th:if="${teamKey != null}">
+								<button
+									type="button"
+									class="text-xs uppercase tracking-wider text-white/45 hover:text-[#E7FF02] transition-colors select-none"
+									style="font-family: 'Work Sans', sans-serif; font-weight: 700;"
+									data-workspace-dropdown-toggle="team"
+									aria-haspopup="true"
+									aria-controls="workspaceTeamCodeDropdown"
+									aria-expanded="false"
+									id="workspaceTeamCodeToggle"
+								>
+									Team code <span aria-hidden="true" class="text-white/45"></span>
+								</button>
+								<div
+									id="workspaceTeamCodeDropdown"
+									class="hidden absolute right-0 mt-2 z-50 w-64 border border-white/20 bg-black px-4 py-3 text-left"
+									role="dialog"
+									aria-label="Team code"
+									aria-hidden="true"
+								>
+									<p class="text-xs uppercase tracking-wider text-white/45" style="font-family: 'Work Sans', sans-serif; font-weight: 700;">Team code</p>
+									<div class="mt-2 flex items-center justify-between gap-2">
+										<div
+											id="workspaceTeamCodeValue"
+											class="font-mono text-sm text-white/80 tracking-widest cursor-pointer select-text transition-colors hover:text-white"
+											th:text="${teamKey}"
+											title="Copy to clipboard"
+											role="button"
+											tabindex="0"
+										>
+											TEAMCODE
+										</div>
+										<button
+											type="button"
+											class="shrink-0 p-1 border border-white/20 hover:border-[#E7FF02] hover:text-[#E7FF02] transition-colors"
+											aria-label="Copy team code"
+											title="Copy to clipboard"
+											data-workspace-copy-team-code
+										>
+											<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="1.8" class="w-4 h-4">
+												<rect x="9" y="9" width="10" height="10" rx="1"></rect>
+												<rect x="5" y="5" width="10" height="10" rx="1"></rect>
+											</svg>
+										</button>
+									</div>
+									<div class="text-xs text-white/45 mt-2" id="workspaceCopyTeamCodeStatus" aria-live="polite"></div>
+								</div>
+							</div>
+						</div>
+						<form th:action="@{/logout}" method="post">
+							<input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}" />
+							<button type="submit" class="px-4 py-2 text-black font-bold hover:opacity-80 transition-opacity text-sm" style="background-color: #E7FF02;">
+								Logout
+							</button>
+						</form>
+					</div>
+				</div>
+			</div>
+		</header>
+	</th:block>
+</body>
+</html>
