+++ src/main/resources/templates/workspace/_main.html
@@ -1,0 +1,91 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<body>
+	<th:block th:fragment="mainContent">
+		<main class="w-full px-0 sm:px-0">
+			<div class="px-6 sm:px-8 py-6 border-b border-white/10">
+				<div class="flex items-center justify-between gap-6">
+					<div class="flex items-start gap-3 min-w-0">
+						<div id="wsSidebarToggleCollapsedHost" class="hidden shrink-0"></div>
+						<div class="min-w-0">
+							<h1 class="text-base uppercase tracking-wider text-white/70" style="font-family: 'Work Sans', sans-serif; font-weight: 700;">Test cases</h1>
+							<p class="text-xs text-white/45 mt-1">Click row body to select. Use Preview/Edit actions for drawer details.</p>
+						</div>
+					</div>
+					<div class="text-right">
+						<p class="text-xs text-white/45">Total</p>
+						<p id="wsTotalCount" class="text-sm text-white/70" th:text="${testCaseCount != null ? testCaseCount : '-'}">-</p>
+					</div>
+				</div>
+			</div>
+
+			<div class="flex min-h-[calc(100vh-12rem)]">
+				<aside id="wsSidebar" class="w-[18rem] sm:w-[20rem] border-r border-white/10 bg-black shrink-0">
+					<div id="wsSidebarInner" class="px-6 py-5">
+						<div id="wsSidebarHeader" class="flex items-center justify-between gap-3">
+							<p id="wsSidebarTitle" class="text-xs uppercase tracking-wider text-white/45" style="font-family: 'Work Sans', sans-serif; font-weight: 700;">Directory</p>
+							<div id="wsSidebarToggleExpandedHost" class="shrink-0">
+								<button
+									id="wsSidebarToggle"
+									type="button"
+									class="shrink-0 h-10 w-10 flex items-center justify-center border border-white/20 hover:border-[#E7FF02] hover:text-[#E7FF02] transition-colors text-white/70"
+									aria-label="Collapse repository sidebar"
+									aria-controls="wsSidebar"
+									aria-expanded="true"
+								>
+									<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" class="w-4 h-4">
+										<path d="M15 6l-6 6 6 6" />
+									</svg>
+								</button>
+							</div>
+						</div>
+						<div id="wsSidebarContent" class="mt-4" aria-label="Folder tree">
+							<div id="wsFolderLoading" class="text-xs text-white/45">Loading folders...</div>
+							<div id="wsFolderEmpty" class="hidden text-xs text-white/45">No folders found.</div>
+							<nav id="wsFolderTree" class="mt-2 space-y-1"></nav>
+						</div>
+					</div>
+				</aside>
+
+				<section class="flex-1 min-w-0">
+					<div class="overflow-x-hidden">
+						<table class="w-full table-fixed text-sm">
+							<thead class="text-white/50">
+								<tr class="border-b border-white/10">
+									<th class="px-2 sm:px-4 py-3 w-10"></th>
+									<th class="px-2 sm:px-3 py-3 w-10">
+										<input id="wsSelectAll" type="checkbox" class="h-4 w-4 accent-[#E7FF02]" aria-label="Select all" />
+									</th>
+									<th class="text-left px-3 sm:px-6 py-3 font-medium">Title</th>
+									<th class="text-left px-3 sm:px-6 py-3 font-medium w-28">Status</th>
+									<th class="text-left px-3 sm:px-6 py-3 font-medium w-44">Tags</th>
+									<th class="text-left pl-3 pr-6 sm:pl-6 sm:pr-4 py-3 font-medium w-44">Last updated</th>
+									<th class="text-right pl-0 pr-20 sm:pr-20 py-3 font-medium w-44">Actions</th>
+								</tr>
+							</thead>
+							<tbody id="wsTbody" class="text-white/75">
+								<tr id="wsLoadingRow" class="border-b border-white/10">
+									<td class="px-6 sm:px-8 py-6" colspan="7">
+										<p class="text-sm text-white/45">Loading test cases...</p>
+									</td>
+								</tr>
+							</tbody>
+						</table>
+					</div>
+				</section>
+			</div>
+			<div class="px-6 sm:px-8 py-4 border-t border-white/10 flex items-center justify-between gap-3">
+				<p id="wsPageInfo" class="text-xs text-white/45">Page 1 of 1</p>
+				<div class="flex items-center gap-2">
+					<button id="wsPrevPage" type="button" class="px-3 py-2 border border-white/20 hover:border-[#E7FF02] hover:text-[#E7FF02] transition-colors text-xs disabled:opacity-40 disabled:cursor-not-allowed">
+						Previous
+					</button>
+					<button id="wsNextPage" type="button" class="px-3 py-2 border border-white/20 hover:border-[#E7FF02] hover:text-[#E7FF02] transition-colors text-xs disabled:opacity-40 disabled:cursor-not-allowed">
+						Next
+					</button>
+				</div>
+			</div>
+		</main>
+	</th:block>
+</body>
+</html>
