+++ src/main/resources/templates/workspace/_notices.html
@@ -1,0 +1,51 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<body>
+	<th:block th:fragment="notices">
+		<div id="workspaceFlashContainer" class="px-6 sm:px-8 pt-5 space-y-3" th:if="${importSuccessMessage != null or importErrorMessage != null}">
+			<div data-flash-item th:if="${importSuccessMessage != null}" class="border border-white/10 bg-black/95 px-4 py-3 text-sm text-white/80" style="border-color: rgba(231, 255, 2, 0.35);">
+				<div class="flex items-start justify-between gap-3">
+					<div class="min-w-0">
+						<span class="font-bold text-black px-2 py-1 mr-3" style="background-color: #E7FF02;">OK</span>
+						<span th:text="${importSuccessMessage}">Upload applied.</span>
+					</div>
+					<button type="button"
+						aria-label="Dismiss notification"
+						class="shrink-0 px-2 py-1 border border-white/20 hover:border-[#E7FF02] hover:text-[#E7FF02] transition-colors text-xs leading-none"
+						data-workspace-flash-dismiss>
+						x
+					</button>
+				</div>
+			</div>
+			<div data-flash-item th:if="${importErrorMessage != null}" class="border border-white/10 bg-black/95 px-4 py-3 text-sm text-white/80">
+				<div class="flex items-start justify-between gap-3">
+					<div class="min-w-0">
+						<span class="font-bold text-black px-2 py-1 mr-3 bg-white/70">Error</span>
+						<span th:text="${importErrorMessage}">Import failed.</span>
+					</div>
+					<button type="button"
+						aria-label="Dismiss notification"
+						class="shrink-0 px-2 py-1 border border-white/20 hover:border-[#E7FF02] hover:text-[#E7FF02] transition-colors text-xs leading-none"
+						data-workspace-flash-dismiss>
+						x
+					</button>
+				</div>
+			</div>
+		</div>
+
+		<div id="importNoticeContainer" class="px-6 sm:px-8 pt-5 hidden" aria-live="polite">
+			<div id="importNotice" class="border border-white/10 bg-black/95 px-4 py-3 text-sm text-white/80">
+				<div class="flex items-start justify-between gap-3">
+					<div class="min-w-0">
+						<span id="importNoticeBadge" class="font-bold text-black px-2 py-1 mr-3 bg-white/70">Error</span>
+						<span id="importNoticeMessage" class="whitespace-pre-line">Import status.</span>
+					</div>
+					<button id="importNoticeClose" type="button" aria-label="Dismiss notification" class="shrink-0 px-2 py-1 border border-white/20 hover:border-[#E7FF02] hover:text-[#E7FF02] transition-colors text-xs leading-none">
+						x
+					</button>
+				</div>
+			</div>
+		</div>
+	</th:block>
+</body>
+</html>
