+++ src/main/resources/templates/workspace/_organize-modal.html
@@ -1,0 +1,47 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<body>
+	<th:block th:fragment="organizeModal">
+		<div id="organizeModal" class="fixed inset-0 z-50 hidden" aria-hidden="true">
+			<div id="organizeBackdrop" class="absolute inset-0 bg-black/75 backdrop-blur-sm"></div>
+			<div class="absolute inset-0 flex items-center justify-center p-4 sm:p-6">
+				<div id="organizePanel" class="w-full max-w-xl border border-[#E7FF02]/30 bg-black/95 shadow-[0_0_0_1px_rgba(231,255,2,0.08),0_30px_80px_rgba(0,0,0,0.75)] p-6 sm:p-7 transition-all duration-150 translate-y-3 opacity-0 rounded-xl">
+					<div class="flex items-start justify-between gap-4">
+						<div class="min-w-0">
+							<p class="text-[11px] uppercase tracking-[0.18em] text-[#E7FF02]/80" style="font-family: 'Work Sans', sans-serif; font-weight: 700;">Organize Selection</p>
+							<h3 class="text-xl font-semibold text-white mt-2 leading-tight">Move selected test cases</h3>
+							<p class="text-xs text-white/55 mt-2">Choose an existing folder or type a new target path.</p>
+						</div>
+						<button id="organizeClose" type="button" class="px-3 py-2 border border-white/25 hover:border-[#E7FF02] hover:text-[#E7FF02] transition-colors text-xs uppercase tracking-wider text-white/70">
+							Close
+						</button>
+					</div>
+
+					<div class="mt-6 space-y-5">
+						<div>
+							<label for="organizeFolderSelect" class="block text-[11px] uppercase tracking-[0.14em] text-white/50" style="font-family: 'Work Sans', sans-serif; font-weight: 700;">Existing folder</label>
+							<select id="organizeFolderSelect" class="mt-2 w-full bg-black border border-white/20 px-3 py-3.5 text-sm text-white/90 focus:outline-none focus:border-[#E7FF02] focus:ring-2 focus:ring-[#E7FF02]/20">
+								<option value="">Select a folder</option>
+							</select>
+						</div>
+						<div>
+							<label for="organizeFolderInput" class="block text-[11px] uppercase tracking-[0.14em] text-white/50" style="font-family: 'Work Sans', sans-serif; font-weight: 700;">Target folder path</label>
+							<input id="organizeFolderInput" type="text" placeholder="e.g. regression/mobile/smoke" class="mt-2 w-full bg-black border border-white/20 px-4 py-3.5 text-sm text-white/90 placeholder:text-white/35 focus:outline-none focus:border-[#E7FF02] focus:ring-2 focus:ring-[#E7FF02]/20" />
+							<p id="organizeError" class="hidden mt-2 text-xs text-rose-300">A target folder is required.</p>
+						</div>
+					</div>
+
+					<div class="mt-7 flex items-center justify-end gap-3 border-t border-white/10 pt-5">
+						<button id="organizeCancel" type="button" class="px-4 py-2.5 border border-white/20 hover:border-[#E7FF02] hover:text-[#E7FF02] transition-colors text-sm text-white/80">
+							Cancel
+						</button>
+						<button id="organizeSave" type="button" class="px-5 py-2.5 text-black font-bold hover:opacity-85 transition-opacity text-sm" style="background-color: #E7FF02;">
+							Save
+						</button>
+					</div>
+				</div>
+			</div>
+		</div>
+	</th:block>
+</body>
+</html>
